backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
board==1.0
certifi==2021.10.8
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==4.0.3
dnspython==2.6.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==2.5.1
fqdn==1.5.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.3
ipykernel==5.5.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.10.0
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.5.8
notebook==6.5.6
notebook_shim==0.2.4
openai==1.3.7
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pdfdocument==4.0.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pymongo==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
reportlab==4.0.8
requests==2.27.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.36
sqlparse==0.4.2
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==1.26.9
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
